Temporarily removed the broken sell command until I am able to fix it.

+++ src/main/java/com/github/carthax08/simplecurrencies/PapiExpansion/SimpleCurrenciesExpansion.java
@@ -87,7 +87,7 @@ public class SimpleCurrenciesExpansion extends PlaceholderExpansion {
             return "";
         }
         if (plugin.getConfig().getBoolean("currencies." + identifier + ".enabled")) {
-            return String.valueOf(plugin.getConfig().getInt("players." + player.getUniqueId().toString() + "." + identifier));
+            return String.valueOf(plugin.getConfig().getDouble("players." + player.getUniqueId().toString() + "." + identifier));
         }
         return null;
     }
